pyproject.toml
[tool.poetry]
name = "financeiroapi"
version = "0.1.0"
description = ""
authors = ["sebasthub <playandroidhb@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
sentry-sdk = {extras = ["fastapi"], version = "^2.3.1"}
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
pydantic-settings = "^2.3.4"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
